feat: registration by phone, before confirm implementation
