--- my-project-OLD2/src/admin/app.example.js
@@ -1,39 +1,0 @@
-const config = {
-  locales: [
-    // 'ar',
-    // 'fr',
-    // 'cs',
-    // 'de',
-    // 'dk',
-    // 'es',
-    // 'he',
-    // 'id',
-    // 'it',
-    // 'ja',
-    // 'ko',
-    // 'ms',
-    // 'nl',
-    // 'no',
-    // 'pl',
-    // 'pt-BR',
-    // 'pt',
-    // 'ru',
-    // 'sk',
-    // 'sv',
-    // 'th',
-    // 'tr',
-    // 'uk',
-    // 'vi',
-    // 'zh-Hans',
-    // 'zh',
-  ],
-};
-
-const bootstrap = (app) => {
-  console.log(app);
-};
-
-export default {
-  config,
-  bootstrap,
-};
