--- my-project/.eslintignore
@@ -1,3 +1,0 @@
-.cache
-build
-**/node_modules/**
