--- my-project/config/api.js
@@ -1,7 +1,0 @@
-module.exports = {
-  rest: {
-    defaultLimit: 25,
-    maxLimit: 100,
-    withCount: true,
-  },
-};
