--- my-project/public/robots.txt
@@ -1,3 +1,0 @@
-# To prevent search engines from seeing the site altogether, uncomment the next two lines:
-# User-Agent: *
-# Disallow: /
