--- my-project/src/admin/webpack.config.example.js
@@ -1,9 +1,0 @@
-'use strict';
-
-/* eslint-disable no-unused-vars */
-module.exports = (config, webpack) => {
-  // Note: we provide webpack above so you should not `require` it
-  // Perform customizations to webpack config
-  // Important: return the modified config
-  return config;
-};
