+++ plugins/sms/server/services/sms.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = ({ strapi }) => ({
+  async send({ to, text }) {
+    return true;
+  },
+});
