+++ src/admin/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "@strapi/typescript-utils/tsconfigs/admin",
   "include": [
-    "../plugins/**/admin/src/**/*",
+    "../plugins/**/admin/src/**/**/*",
     "./"
   ],
   "exclude": [
