--- src/plugins/sms/services/sms.ts
@@ -1,5 +1,0 @@
-export default () => ({
-  async send({ to, text }: { to: string; text: string }) {
-    console.log(`SMS to ${to}: ${text}`);
-  },
-});
