--- src/plugins/sms/strapi-server.ts
@@ -1,9 +1,0 @@
-import sms from './services/sms';
-
-export default () => ({
-  register() {},
-  bootstrap() {},
-  services: {
-    sms,
-  },
-});
