+++ tsconfig.json
@@ -2,22 +2,22 @@
   "extends": "@strapi/typescript-utils/tsconfigs/server",
   "compilerOptions": {
     "outDir": "dist",
-    "rootDir": "."
+    "rootDir": ".",
+    "allowJs": true
   },
   "include": [
-    "./",
     "./**/*.ts",
     "./**/*.js",
-    "src/**/*.json"
+    "src/**/*.json",
+    "plugins/**/*"
   ],
   "exclude": [
-    "node_modules/",
-    "build/",
-    "dist/",
-    ".cache/",
-    ".tmp/",
-    "src/admin/",
-    "**/*.test.*",
-    "src/plugins/**"
+    "node_modules",
+    "build",
+    "dist",
+    ".cache",
+    ".tmp",
+    "src/admin",
+    "**/*.test.*"
   ]
 }
